Add communication between boards
The communication (refresh) is simply activated by using "_active_%s" in the
domain of the act_window (where %s is the id of the other act_window).  When
the selection of one board is changed, every other board domain are evaluated
and if they change the corresponding board is reloaded.