Fix gtk.Button state corruption
By default gtk.Button grab the focus on button-press, this can potentially
trigger a RPC request from by example a Many2One filled with search text. If
the request takes too much time, RPCProgress starts to process events and this
put the button in a corrupted state (it is displayed as pressed when mouse if
over).
To avoid, we set the value of the view when mouse move over a button or outside
the form.

review447001